+++ civipy/event.py
@@ -1,7 +1,9 @@
 from civipy.base import CiviCRMBase
 
+
 class CiviEvent(CiviCRMBase):
     pass
+
 
 class CiviParticipant(CiviCRMBase):
     pass