+++ civipy/order.py
@@ -1,9 +1,11 @@
 from civipy.base import CiviCRMBase
 
+
 class CiviOrder(CiviCRMBase):
     pass
+
 
 class CiviPayment(CiviCRMBase):
     @classmethod
     def cancel(cls, **kwargs):
-        return cls._post_method()('cancel', "Payment", kwargs)
+        return cls._post_method()("cancel", "Payment", kwargs)